requirements.txt
cycler==0.10.0
matplotlib==2.0.2
numpy==1.12.1
olefile==0.44
Pillow==4.1.1
Pmw==2.0.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
timezonefinder==2.1.0
